refactor(llm): 完成 handler.py 拆分（第二步）——创建 _commands.py，重写 handler.py
- 新建 packages/llm/_commands.py（396 行）：提取全部 _cmd_* 命令方法
  (_cmd_help, _cmd_sid, _cmd_reset, _cmd_llm, _cmd_status,
   _cmd_persona, _cmd_list_manage, _cmd_llm_group, _cmd_model_list)
  组成 _CommandsMixin，跨 mixin 调用使用 # type: ignore[attr-defined]

- 重写 packages/llm/handler.py（297 行）：
  LLMHandler 改为继承所有 6 个 mixin
  (_CommandsMixin, _GroupMixin, _WakeMixin,
   _ContextMixin, _HistoryMixin, _HelpersMixin)
  保留 __init__、_get_conv_lock（per-key asyncio.Lock 缓存）、
  handle()、_handle_message()、_run_tool_loop()
  删除全部原有单体代码，彻底消除新旧代码混合
